+++ llvm/include/llvm/Debuginfod/Debuginfod.h
@@ -22,9 +22,9 @@
 
 #include "llvm/ADT/StringMap.h"
 #include "llvm/ADT/StringRef.h"
+#include "llvm/HTTP/HTTPServer.h"
 #include "llvm/Object/BuildID.h"
 #include "llvm/Support/Error.h"
-#include "llvm/Support/HTTP/HTTPServer.h"
 #include "llvm/Support/MemoryBuffer.h"
 #include "llvm/Support/Mutex.h"
 #include "llvm/Support/RWMutex.h"
