+++ llvm/include/llvm/HTTP/HTTPServer.h
@@ -13,8 +13,8 @@
 ///
 //===----------------------------------------------------------------------===//
 
-#ifndef LLVM_SUPPORT_HTTP_HTTPSERVER_H
-#define LLVM_SUPPORT_HTTP_HTTPSERVER_H
+#ifndef LLVM_HTTP_HTTPSERVER_H
+#define LLVM_HTTP_HTTPSERVER_H
 
 #include "llvm/ADT/StringRef.h"
 #include "llvm/Support/Error.h"
@@ -130,4 +130,4 @@ public:
 };
 } // end namespace llvm
 
-#endif // LLVM_SUPPORT_HTTP_HTTPSERVER_H
+#endif // LLVM_HTTP_HTTPSERVER_H
