+++ llvm/lib/CMakeLists.txt
@@ -33,6 +33,7 @@ add_subdirectory(ObjectYAML)
 add_subdirectory(Option)
 add_subdirectory(Remarks)
 add_subdirectory(RemoteCachingService)
+add_subdirectory(HTTP)
 add_subdirectory(Debuginfod)
 add_subdirectory(DebugInfo)
 add_subdirectory(DWARFCFIChecker)
