+++ llvm/lib/Support/CMakeLists.txt
@@ -135,7 +135,6 @@ if (UNIX AND ${CMAKE_SYSTEM_NAME} MATCHES "AIX")
 endif()
 
 add_subdirectory(BLAKE3)
-add_subdirectory(HTTP)
 
 add_llvm_component_library(LLVMSupport
   ABIBreak.cpp
