+++ llvm/tools/llvm-cov/CMakeLists.txt
@@ -1,7 +1,6 @@
 set(LLVM_LINK_COMPONENTS
   Core
   Support
-  SupportHTTP
   Object
   Coverage
   ProfileData
@@ -23,4 +22,4 @@ add_llvm_tool(llvm-cov
   TestingSupport.cpp
   )
 
-target_link_libraries(llvm-cov PRIVATE LLVMDebuginfod)
+target_link_libraries(llvm-cov PRIVATE LLVMHTTP LLVMDebuginfod)
