+++ llvm/tools/llvm-debuginfod-find/CMakeLists.txt
@@ -2,7 +2,6 @@ set(LLVM_LINK_COMPONENTS
   Option
   Object
   Support
-  SupportHTTP
   )
 set(LLVM_TARGET_DEFINITIONS Opts.td)
 tablegen(LLVM Opts.inc -gen-opt-parser-defs)
@@ -14,8 +13,9 @@ add_llvm_tool(llvm-debuginfod-find
   DebugInfodFindOptsTableGen
   GENERATE_DRIVER
   )
+
 if(NOT LLVM_TOOL_LLVM_DRIVER_BUILD)
-  target_link_libraries(llvm-debuginfod-find PRIVATE LLVMDebuginfod)
+  target_link_libraries(llvm-debuginfod-find PRIVATE LLVMDebuginfod LLVMHTTP)
 endif()
 if(LLVM_INSTALL_BINUTILS_SYMLINKS)
   add_llvm_tool_symlink(debuginfod-find llvm-debuginfod-find)
