+++ llvm/tools/llvm-debuginfod/CMakeLists.txt
@@ -1,7 +1,6 @@
 set(LLVM_LINK_COMPONENTS
   Option
   Support
-  SupportHTTP
   )
 set(LLVM_TARGET_DEFINITIONS Opts.td)
 tablegen(LLVM Opts.inc -gen-opt-parser-defs)
@@ -16,7 +15,7 @@ add_llvm_tool(llvm-debuginfod
   )
 
 if(NOT LLVM_TOOL_LLVM_DRIVER_BUILD)
-  target_link_libraries(llvm-debuginfod PRIVATE LLVMDebuginfod)
+  target_link_libraries(llvm-debuginfod PRIVATE LLVMDebuginfod LLVMHTTP)
 endif()
 
 if(LLVM_INSTALL_BINUTILS_SYMLINKS)
