+++ llvm/tools/llvm-profdata/CMakeLists.txt
@@ -3,7 +3,6 @@ set(LLVM_LINK_COMPONENTS
   Object
   ProfileData
   Support
-  SupportHTTP
   )
 
 add_llvm_tool(llvm-profdata
