+++ llvm/unittests/CMakeLists.txt
@@ -45,6 +45,7 @@ add_subdirectory(ExecutionEngine)
 add_subdirectory(FileCheck)
 add_subdirectory(Frontend)
 add_subdirectory(FuzzMutate)
+add_subdirectory(HTTP)
 add_subdirectory(InterfaceStub)
 add_subdirectory(IR)
 add_subdirectory(LineEditor)
