+++ llvm/unittests/Debuginfod/CMakeLists.txt
@@ -1,7 +1,3 @@
-set(LLVM_LINK_COMPONENTS
-  SupportHTTP
-  )
-
 add_llvm_unittest(DebuginfodTests
   DebuginfodTests.cpp
   )
