+++ llvm/unittests/HTTP/CMakeLists.txt
@@ -1,0 +1,8 @@
+add_llvm_unittest(HTTPTests
+  HTTPServerTests.cpp
+  )
+
+target_link_libraries(HTTPTests PRIVATE
+  LLVMHTTP
+  LLVMTestingSupport
+  )
