+++ llvm/unittests/HTTP/HTTPServerTests.cpp
@@ -7,9 +7,9 @@
 //===----------------------------------------------------------------------===//
 
 #include "llvm/ADT/StringExtras.h"
+#include "llvm/HTTP/HTTPClient.h"
+#include "llvm/HTTP/HTTPServer.h"
 #include "llvm/Support/Error.h"
-#include "llvm/Support/HTTP/HTTPClient.h"
-#include "llvm/Support/HTTP/HTTPServer.h"
 #include "llvm/Support/ThreadPool.h"
 #include "llvm/Testing/Support/Error.h"
 #include "gmock/gmock.h"
